@@ -13,7 +13,7 @@
     }
   ],
   "require": {
-    "php": ">=8.0",
+    "php": ">=8.2",
     "fond-of-oryx/erp-order": "^4.4.0",
     "fond-of-oryx/erp-order-page-search-extension": "^1.0.0",
     "fond-of-oryx/erp-order-permission": "^1.0.0 || ^2.0.0",
@@ -42,7 +42,7 @@
     "psr-4": {
       "FondOfOryx\\": "tests/FondOfOryx/",
       "Generated\\": "src/Generated/",
-      "Orm\\Zed\\": "src/Orm/Zed/"
+      "Orm\\": "src/Orm/"
     }
   },
   "config": {
